+++ models/datasources.py
@@ -147,15 +147,13 @@ class DataSourceLayerManager(object):
         return [self._line_layer, self._node_layer, self._pumpline_layer]
 
     def _get_result_layers_groundwater(self):
-        lines_path = os.path.join(self.datasource_dir, 'flowlines.sqlite')
-        nodes_path = os.path.join(self.datasource_dir, 'nodes.sqlite')
-        pumps_path = os.path.join(self.datasource_dir, 'pumplines.sqlite')
+        sqlite_path = os.path.join(self.datasource_dir, 'gridadmin.sqlite')
         self._line_layer = self._line_layer or get_or_create_flowline_layer(
-            self.datasource, lines_path)
+            self.datasource, sqlite_path)
         self._node_layer = self._node_layer or get_or_create_node_layer(
-            self.datasource, nodes_path)
+            self.datasource, sqlite_path)
         self._pumpline_layer = self._pumpline_layer or \
-            get_or_create_pumpline_layer(self.datasource, pumps_path)
+            get_or_create_pumpline_layer(self.datasource, sqlite_path)
         return [self._line_layer, self._node_layer, self._pumpline_layer]