+++ utils/layer_from_netCDF.py
@@ -22,48 +22,68 @@ PUMPLINES_LAYER_NAME = 'pumplines'
 IGNORE_FIRST = slice(1, None, None)
 
 
+def contains_layer(sqlite_path, layer_name):
+    driver = ogr.GetDriverByName('SQLite')
+    data_source = driver.Open(sqlite_path)
+    has_layer = False
+    for i in range(data_source.GetLayerCount()):
+        lyr = data_source.GetLayer(i)
+        if lyr.GetName() == layer_name:
+            has_layer = True
+            break
+    data_source = None  # close data source
+    return has_layer
+
+
+def _get_vec_lyr(sqlite_path, layer_name):
+    """Helper function to construct a QgsVectorLayer."""
+    uri = QgsDataSourceURI()
+    uri.setDatabase(sqlite_path)
+    uri.setDataSource('', layer_name, 'geometry')
+    return QgsVectorLayer(uri.uri(), layer_name, 'spatialite')
+
+
 @disable_sqlite_synchronous
 def get_or_create_flowline_layer(ds, output_path):
-    if not os.path.exists(output_path):
+    if not os.path.exists(output_path) or not \
+            contains_layer(output_path, FLOWLINES_LAYER_NAME):
         ga = ds.gridadmin  # TODO: to implement
         from .gridadmin import QgisLinesOgrExporter
         exporter = QgisLinesOgrExporter('dont matter')
         exporter.driver = ogr.GetDriverByName('SQLite')
-        sliced = ga.lines.slice(IGNORE_FIRST).data
-        exporter.save(output_path, sliced.data, sliced.epsg_code)
-    uri = QgsDataSourceURI()
-    uri.setDatabase(output_path)
-    uri.setDataSource('', FLOWLINES_LAYER_NAME, 'geometry')
-    return QgsVectorLayer(uri.uri(), FLOWLINES_LAYER_NAME, 'spatialite')
+        sliced = ga.lines.slice(IGNORE_FIRST)
+        exporter.save(
+            output_path, FLOWLINES_LAYER_NAME, sliced.data, sliced.epsg_code)
+    return _get_vec_lyr(output_path, FLOWLINES_LAYER_NAME)
 
 
 @disable_sqlite_synchronous
 def get_or_create_node_layer(ds, output_path):
-    if not os.path.exists(output_path):
+    if not os.path.exists(output_path) or not \
+            contains_layer(output_path, NODES_LAYER_NAME):
         ga = ds.gridadmin  # TODO: to implement
         from .gridadmin import QgisNodesOgrExporter
         exporter = QgisNodesOgrExporter('dont matter')
         exporter.driver = ogr.GetDriverByName('SQLite')
-        sliced = ga.nodes.slice(IGNORE_FIRST).data
-        exporter.save(output_path, sliced.data, sliced.epsg_code)
-    uri = QgsDataSourceURI()
-    uri.setDatabase(output_path)
-    uri.setDataSource('', NODES_LAYER_NAME, 'geometry')
-    return QgsVectorLayer(uri.uri(), NODES_LAYER_NAME, 'spatialite')
+        sliced = ga.nodes.slice(IGNORE_FIRST)
+        exporter.save(
+            output_path, NODES_LAYER_NAME, sliced.data, sliced.epsg_code)
+    return _get_vec_lyr(output_path, NODES_LAYER_NAME)
 
 
 @disable_sqlite_synchronous
 def get_or_create_pumpline_layer(ds, output_path):
-    if not os.path.exists(output_path):
+    if not os.path.exists(output_path) or not \
+            contains_layer(output_path, PUMPLINES_LAYER_NAME):
         ga = ds.gridadmin  # TODO: to implement
         from .gridadmin import QgisPumpsOgrExporter
         exporter = QgisPumpsOgrExporter(node_data=ga.nodes.data)
         exporter.driver = ogr.GetDriverByName('SQLite')
-        exporter.save(output_path, ga.pumps.data, ga.pumps.epsg_code)
-    uri = QgsDataSourceURI()
-    uri.setDatabase(output_path)
-    uri.setDataSource('', PUMPLINES_LAYER_NAME, 'geometry')
-    return QgsVectorLayer(uri.uri(), PUMPLINES_LAYER_NAME, 'spatialite')
+        exporter.save(
+            output_path, PUMPLINES_LAYER_NAME, ga.pumps.data,
+            ga.pumps.epsg_code
+        )
+    return _get_vec_lyr(output_path, PUMPLINES_LAYER_NAME)
 
 
 def make_flowline_layer(ds, spatialite, progress_bar=None):